Refactor CUDA and data handling in NCT
- Updated Device::info to remove unnecessary line.
- Refactored FFT functions to use u32 instead of size_t for dimensions and batch parameters.
- Changed operator overloads in DcmTag and DcmVR to be const and moved public access specifiers.
- Simplified MduTbl constructor and destructor, using default implementations.
- Modified RawData structure to use more descriptive member names and removed unnecessary member variables.
- Updated FFT functions to use u32 for dimensions and batch parameters.
- Changed NdArray and NdView templates to use u32 instead of usize for better clarity.
- Refactored cone_bp and cone_fp algorithms to improve readability and maintainability.
- Adjusted FDK filter creation to use updated parameter access.
- Updated command-line argument handling in tools for better clarity and consistency.

+++ src/nct/math/ndarray.h
@@ -7,7 +7,7 @@ namespace nct::math {
 
 using cuda::Alloc;
 
-template <class T, usize N = 1, class A = Alloc>
+template <class T, u32 N = 1, class A = Alloc>
 class NdArray {
   NdView<T, N> _view{};
   A _a{};
@@ -22,8 +22,8 @@ class NdArray {
   NdArray(const NdArray&) = delete;
   NdArray& operator=(const NdArray&) = delete;
 
-  static auto with_shape(const usize (&dims)[N], A a = A{}) -> NdArray {
-    usize step[N] = {1U};
+  static auto with_shape(const u32 (&dims)[N], A a = A{}) -> NdArray {
+    u32 step[N] = {1U};
     for (auto i = 0U; i < N; ++i) {
       step[i] = i == 0 ? 1 : step[i - 1] * dims[i - 1];
     }
@@ -46,11 +46,11 @@ class NdArray {
     return _view._data;
   }
 
-  auto size() const -> const usize (&)[N] {
+  auto size() const -> const u32 (&)[N] {
     return _view._size;
   }
 
-  auto numel() const -> usize {
+  auto numel() const -> u32 {
     return _view.numel();
   }
 
@@ -65,19 +65,19 @@ class NdArray {
   }
 
  public:
-  inline auto operator[](usize idx) const -> T {
+  inline auto operator[](u32 idx) const -> T {
     return _view._data[idx];
   }
 
-  inline auto operator[](usize idx) -> T& {
+  inline auto operator[](u32 idx) -> T& {
     return _view._data[idx];
   }
 
-  inline auto operator[](const usize (&idxs)[N]) const -> T {
+  inline auto operator[](const u32 (&idxs)[N]) const -> T {
     return _view[idxs];
   }
 
-  inline auto operator[](const usize (&idxs)[N]) -> T& {
+  inline auto operator[](const u32 (&idxs)[N]) -> T& {
     return _view[idxs];
   }
 
@@ -103,7 +103,7 @@ class NdArray {
 #endif
 };
 
-template <class T, usize N>
+template <class T, u32 N>
 void fill(NdView<T, N> view, T val) {
   const auto n = view.numel();
   if (val == 0) {
